Fix Coverity 1491741 and 1491742.

+++ src/sys_tree.c
@@ -239,7 +239,7 @@ void sys_tree__update(void)
 			&& db.now_s - db.config->sys_interval >= last_update){
 
 		uptime = db.now_s - start_time;
-		len = (uint32_t)snprintf(buf, BUFLEN, "%d seconds", (int)uptime);
+		len = (uint32_t)snprintf(buf, BUFLEN, "%" PRIu64 " seconds", (uint64_t)uptime);
 		db__messages_easy_queue(NULL, "$SYS/broker/uptime", SYS_TREE_QOS, len, buf, 1, 60, NULL);
 
 		sys_tree__update_clients(buf);
